requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
python-dotenv
sqlalchemy
databases  
sentence-transformers
matplotlib